requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
anthropic==0.39.0
supabase==2.9.0
redis[hiredis]==5.1.0
httpx==0.27.0
tenacity==9.0.0
elevenlabs==1.9.0
google-api-python-client==2.150.0
google-auth==2.35.0
twilio==9.3.0
